ARM: dts: imx6-sabrelite: add video camera devices and connections
Enables the host video camera interface device as a child of media0,
and the OV5642 parallel-bus sensor. The host camif0 device connects to
the OV5642 via the parallel-bus mux input on the ipu1_csi0_mux.

Note there is a pin conflict with GPIO6. This pin functions as a power
input pin to the OV5642, but ENET requires it to wake-up the ARM cores
on normal RX and TX packet done events (see 6261c4c8). So by default,
capture from the OV5642 is disabled, enable by uncommenting __ENABLE_OV5642__
macro. Ethernet will still work just not quite as well.

Signed-off-by: Steve Longerbeam <steve_longerbeam@mentor.com>