ARM: dts: imx6-sabreauto: add video camera devices and connections
Enables the host video camera interface device as a child of media0,
and the ADV7180 decoder sensor. The host camif0 device connects to
the ADV7180 via the parallel-bus mux input on the ipu1_csi0_mux.

On the sabreauto, two analog video inputs are routed to the ADV7180,
composite on Ain1, and composite on Ain3. Those inputs are defined
via inputs and input-names under the ADV7180 node. The ADV7180 power
pin is via max7310_b port expander.

Signed-off-by: Steve Longerbeam <steve_longerbeam@mentor.com>